adding copyright section , adding <em>tag
adding copyright section , instead of <i> tag <em>tag is added because emphasis tag tells the browser the words that are enclosed between it should be stressed or should be emphasized, but the i tag only italicizes the text. so it only styles the text to make it slanted. it's good practice to always use emphasis instead of italic because it conveys more information.